backend/pyproject.toml
[project]
name = "cadency-api"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi",
    "uvicorn[standard]",
    "python-multipart",
    "yt-dlp",
    "librosa",
    "numpy",
    "music21",
    "google-genai",
    "python-dotenv",
    "httpx",
    "pydantic>=2.12.5",
    "sqlalchemy[asyncio]>=2.0",
    "aiosqlite>=0.19",
    "alembic>=1.13",
]

[dependency-groups]
dev = [
    "pytest",
    "httpx",
]

[tool.uv]
package = false
